day 15 part 2 worked before but now it works fast
before this commit, it took ~1hr to run but did get the correct answer
with this commit, that is cut down to less than half a second

also, fixed an issue with neighbor detection bounds being off by one for
east and south neighbors. I'm not quite sure how it was not a problem before.